+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.1.24-SNAPSHOT] - 2022-10-31
+## [0.1.24-SNAPSHOT] - 2022-11-24
 
 ### Added
 
+ - Resource manager for calculating CPU and memory usage
  - Logic for checking internal and external "health" of the connector.
  Internal health: dataApp and Usage Control app reachability, audit file volume usage
  External health: DAPS and Clearing House reachability.
@@ -14,8 +15,15 @@ All notable changes to this project will be documented in this file.
 ```
 application.healthcheck.enabled=false
 application.healthcheck.cron.expression=0 */5 * ? * *
+application.healthcheck.resourcemanager.cron.fixedDelay=10000
 ## Threshold in percentages - from max value
 application.healthcheck.threshold.audit=5
+
+application.healthcheck.threshold.cpu=0.5
+application.healthcheck.threshold.memory=10.5
+application.healthcheck.limit.cpu=1.0
+application.healthcheck.limit.memory=40.5
+
 application.healthcheck.dataapp=
 application.healthcheck.usagecontrol
 ```